backend/requirements.txt
flask
flask-cors
tensorflow
numpy
scikit-learn
Pillow
opencv-python-headless
scikit-image
gunicorn